#5 Task E-Shop/eShop/requirements.txt
Django
django-rest-framework
Pillow
psycopg2-binary
python-decouple
